[GGB-12] interface - implements 구조 제거
